+++ packages/react-utils/src/index.ts
@@ -2,3 +2,5 @@ export * from "./is-server";
 export * from "./use-isomorphic-layout-effect";
 export * from "./use-media-query";
 export * from "./use-presence";
+
+export type * from "./types";