+++ packages/react/src/radio/radio.tsx
@@ -1,24 +1,25 @@
 import { Ripple } from "../ripple";
 import { styles } from "./radio.css";
 import { FocusRing } from "../focus";
-import { forwardRef, useEffect, useImperativeHandle, useMemo, useRef, useState } from "react";
+import { forwardRef, useEffect, useImperativeHandle, useMemo, useRef, useState, type ReactNode } from "react";
+import type { ComposableForwardRefExoticComponent, ForwardRefExoticComponentProps } from "@star4/react-utils";
 
 export namespace Radio {
-  export type Props =
+  export type Props<T extends Value> =
     & {
       /**
        * The current value of the radio button
        */
-      value: Value;
+      value: T;
       /**
        * The current value of the radio button group
        */
-      groupValue: Value;
+      groupValue: T;
       /**
        * Callback that fires when {@link Props.value|value} changes
        * @eventProperty
        */
-      onChange?: (value: Value) => void;
+      onChange?: (value: T) => void;
       /**
        * Prevents the user from interacting with the radio button: it cannot be pressed or focused.
        *
@@ -46,20 +47,49 @@ export namespace Radio {
  *
  * @see {@link https://m3.material.io/components/radio-button/overview|Radio button - Material Design 3}
  *
+ * @typeParam T - the value type of the radio button group
+ *
  * @example
  * Use strings as values
  * ```
- * const [value, setValue] = useState("apple");
- * return (
- *   <div>
- *     <Radio groupValue={value} value="apple" onChange={setValue} />
- *     <Radio groupValue={value} value="banana" onChange={setValue} />
- *     <Radio groupValue={value} value="cucumber" onChange={setValue} />
- *   </div>
- * );
+ * const Example = () => {
+ *   const [value, setValue] = useState("apple");
+ *   return (
+ *     <div>
+ *       <Radio groupValue={value} value="apple" onChange={setValue} />
+ *       <Radio groupValue={value} value="banana" onChange={setValue} />
+ *       <Radio groupValue={value} value="cucumber" onChange={setValue} />
+ *     </div>
+ *   );
+ * }
+ * ```
+ *
+ * @example
+ * Explicitly provide value type
+ * ```tsx
+ * const enum Fruit {
+ *   APPLE,
+ *   APRICOT,
+ *   BANANA,
+ *   CUCUMBER,
+ *   ORANGE,
+ * }
+ *
+ * const Example = () => {
+ *   const [fruit, setFruit] = useState<Fruit>(Fruit.APPLE);
+ *   return (
+ *     <div>
+ *       <Radio<Fruit> groupValue={fruit} value={Fruit.APPLE} onChange={setFruit} />
+ *       <Radio<Fruit> groupValue={fruit} value={Fruit.APRICOT} onChange={setFruit} />
+ *       <Radio<Fruit> groupValue={fruit} value={Fruit.BANANA} onChange={setFruit} />
+ *       <Radio<Fruit> groupValue={fruit} value={Fruit.CUCUMBER} onChange={setFruit} />
+ *       <Radio<Fruit> groupValue={fruit} value={Fruit.ORANGE} onChange={setFruit} />
+ *     </div>
+ *   );
+ * }
  * ```
  */
-export const Radio = forwardRef<Radio.Element, Radio.Props>(
+export const Radio = forwardRef<Radio.Element, Radio.Props<Radio.Value>>(
   (
     {
       groupValue,
@@ -108,4 +138,9 @@ export const Radio = forwardRef<Radio.Element, Radio.Props>(
       </div>
     );
   },
-);
+) as ComposableForwardRefExoticComponent<
+  ExoticProps<Radio.Value>,
+  <T extends Radio.Value>(props: ExoticProps<T>) => ReactNode
+>;
+
+type ExoticProps<T extends Radio.Value> = ForwardRefExoticComponentProps<Radio.Element, Radio.Props<T>>;